Separate the checkbox into its own file
BananaSlugFrontendControl and ColorizerFrontend both use the new class. I left the wrappers in their original classes, thought it made more sense to keep specifying the store properties and listeners there.
