+++ plugins/BananaSlug/BananaSlugFrontendControl.js
@@ -12,85 +12,9 @@
 'use strict';
 
 const React = require('react');
-const immutable = require('immutable');
 
 var decorate = require('../../frontend/decorate');
-
-import type {ControlState} from '../../frontend/types.js';
-
-type Props = {
-  state: any,
-  onChange: (v: ControlState) => void,
-};
-
-type State = StateRecord;
-
-type DefaultProps = {};
-
-const StateRecord = immutable.Record({
-  enabled: false,
-});
-
-class BananaSlugFrontendControl extends React.Component {
-  props: Props;
-  defaultProps: DefaultProps;
-  state: State;
-
-  _defaultState: ControlState;
-  _toggle: (b: boolean) => void;
-
-  constructor(props: Props) {
-    super(props);
-    this._toggle = this._toggle.bind(this);
-    this._defaultState = new StateRecord();
-  }
-
-  componentDidMount(): void {
-    if (!this.props.state !== this._defaultState) {
-      this.props.onChange(this._defaultState);
-    }
-  }
-
-  render() {
-    var state = this.props.state || this._defaultState;
-    return (
-      <div style={styles.container} onClick={this._toggle} tabIndex={0}>
-        <input
-          style={styles.checkbox}
-          type="checkbox"
-          checked={state.enabled}
-          readOnly={true}
-        />
-        <span>Trace React updates</span>
-      </div>
-    );
-  }
-
-  _toggle() {
-    var state = this.props.state || this._defaultState;
-    var nextState = state.merge({
-      enabled: !state.enabled,
-    });
-
-    this.props.onChange(nextState);
-  }
-}
-
-var styles = {
-  checkbox: {
-    pointerEvents: 'none',
-  },
-  container: {
-    WebkitUserSelect: 'none',
-    cursor: 'pointer',
-    display: 'inline-block',
-    fontFamily: 'arial',
-    fontSize: '12px',
-    outline: 'none',
-    userSelect: 'none',
-    margin: '0px 4px',
-  },
-};
+var SettingsCheckbox = require('../../frontend/SettingsCheckbox');
 
 var Wrapped = decorate({
   listeners() {
@@ -99,9 +23,10 @@ var Wrapped = decorate({
   props(store) {
     return {
       state: store.bananaslugState,
+      text: 'Trace React Updates',
       onChange: state => store.changeBananaSlug(state),
     };
   },
-}, BananaSlugFrontendControl);
+}, SettingsCheckbox);
 
 module.exports = Wrapped;
